pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sdgft-ml"
version = "0.1.0"
description = "ML surrogate, inverter & anomaly detector for SDGFT theory"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "David A. Besemer" }]

dependencies = [
    "torch>=2.1",
    "torch-geometric>=2.4",
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "scikit-learn>=1.3",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]
wandb = ["wandb>=0.16"]
symbolic = ["pysr>=0.16"]
full = ["sdgft-ml[dev,wandb,symbolic]"]

[project.scripts]
sdgft-predict = "sdgft_ml.api.predict:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sdgft_ml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 88
